rename createTeam.ts file to createTeamModal
